Move configuration classes from utils to their own package
