+++ src/test/resources/org/datadog/jenkins/plugins/datadog/globalConfiguration.xml
@@ -1,10 +1,10 @@
 <?xml version='1.1' encoding='UTF-8'?>
 <org.datadog.jenkins.plugins.datadog.DatadogGlobalConfiguration plugin="datadog@6.0.3">
-    <datadogClientConfiguration class="org.datadog.jenkins.plugins.datadog.util.config.DatadogApiConfiguration">
+    <datadogClientConfiguration class="org.datadog.jenkins.plugins.datadog.configuration.DatadogApiConfiguration">
         <apiUrl>https://my-api-url.com/api/</apiUrl>
         <logIntakeUrl>https://my-log-intake-url.com/v1/input/</logIntakeUrl>
         <webhookIntakeUrl>https://my-webhook-intake-url.com/api/v2/webhook/</webhookIntakeUrl>
-        <apiKey class="org.datadog.jenkins.plugins.datadog.util.config.DatadogTextApiKey">
+        <apiKey class="org.datadog.jenkins.plugins.datadog.configuration.api.key.DatadogTextApiKey">
             <key>{AQAAABAAAAAwB78atDHwzelG9W0Fw5FRNq0ZUaLh1BwpQPKhvs2u84lxkRPDqeUNoVZel+MKwfyTOjuetnituHYGMdvE9bc3kg==}</key>
         </apiKey>
     </datadogClientConfiguration>
