+++ src/test/resources/org/datadog/jenkins/plugins/datadog/globalConfigurationCredentialsKey.xml
@@ -1,10 +1,10 @@
 <?xml version='1.1' encoding='UTF-8'?>
 <org.datadog.jenkins.plugins.datadog.DatadogGlobalConfiguration plugin="datadog@6.0.3">
-    <datadogClientConfiguration class="org.datadog.jenkins.plugins.datadog.util.config.DatadogApiConfiguration">
+    <datadogClientConfiguration class="org.datadog.jenkins.plugins.datadog.configuration.DatadogApiConfiguration">
         <apiUrl>https://my-api-url.com/api/</apiUrl>
         <logIntakeUrl>https://my-log-intake-url.com/v1/input/</logIntakeUrl>
         <webhookIntakeUrl>https://my-webhook-intake-url.com/api/v2/webhook/</webhookIntakeUrl>
-        <apiKey class="org.datadog.jenkins.plugins.datadog.util.config.DatadogCredentialsApiKey">
+        <apiKey class="org.datadog.jenkins.plugins.datadog.configuration.api.key.DatadogCredentialsApiKey">
             <credentialsId>my-api-key-credentials-id</credentialsId>
         </apiKey>
     </datadogClientConfiguration>
